+++ src/Documentation/Markdown/MarkdownDocumentationWriter.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Josef Pihrt and Contributors. Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Text;
 using DotMarkdown;
 using Microsoft.CodeAnalysis;
 
@@ -10,20 +9,16 @@ namespace Roslynator.Documentation.Markdown
     {
         private readonly MarkdownWriter _writer;
 
-        public MarkdownDocumentationWriter(
-            DocumentationModel documentationModel,
-            DocumentationUrlProvider urlProvider,
-            DocumentationOptions options = null,
-            DocumentationResources resources = null) : base(documentationModel, urlProvider, options, resources)
+        public MarkdownDocumentationWriter(DocumentationContext context, MarkdownWriter writer) : base(context)
         {
-            _writer = MarkdownWriter.Create(new StringBuilder());
+            _writer = writer;
         }
 
-        public override void WriteStartDocument()
+        public override void WriteStartDocument(ISymbol symbol, DocumentationFileKind fileKind)
         {
         }
 
-        public override void WriteEndDocument()
+        public override void WriteEndDocument(ISymbol symbol, DocumentationFileKind fileKind)
         {
         }
 
@@ -135,7 +130,7 @@ namespace Roslynator.Documentation.Markdown
 
         public override void WriteLineBreak()
         {
-            WriteRaw(@"\");
+            WriteRaw("  ");
             WriteLine();
         }