+++ src/Documentation/UrlProviders/MicrosoftDocsUrlProvider.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Josef Pihrt and Contributors. Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Immutable;
-using System.Globalization;
-using System.Linq;
 using System.Text;
 using Microsoft.CodeAnalysis;
 using Roslynator.Text;
@@ -19,125 +17,32 @@ namespace Roslynator.Documentation
 
         public override string Name => "Microsoft Docs";
 
-        internal static ImmutableArray<string> GetFolders(ISymbol symbol)
-        {
-            ImmutableArray<string>.Builder builder = ImmutableArray.CreateBuilder<string>();
-
-            if (symbol.Kind == SymbolKind.Namespace
-                && ((INamespaceSymbol)symbol).IsGlobalNamespace)
-            {
-                builder.Add(WellKnownNames.GlobalNamespaceName);
-            }
-            else if (symbol.Kind == SymbolKind.Method
-                && ((IMethodSymbol)symbol).MethodKind == MethodKind.Constructor)
-            {
-                builder.Add(WellKnownNames.ConstructorName);
-            }
-            else if (symbol.Kind == SymbolKind.Property
-                && ((IPropertySymbol)symbol).IsIndexer)
-            {
-                builder.Add("Item");
-            }
-            else
-            {
-                ISymbol explicitImplementation = symbol.GetFirstExplicitInterfaceImplementation();
-
-                if (explicitImplementation != null)
-                {
-                    string name = explicitImplementation
-                        .ToDisplayParts(DocumentationDisplayFormats.ExplicitImplementationFullName, SymbolDisplayAdditionalMemberOptions.UseItemPropertyName)
-                        .Where(part => part.Kind != SymbolDisplayPartKind.Space)
-                        .Select(part => (part.IsPunctuation()) ? part.WithText("-") : part)
-                        .ToImmutableArray()
-                        .ToDisplayString();
-
-                    builder.Add(name);
-                }
-                else
-                {
-                    int arity = symbol.GetArity();
-
-                    if (arity > 0)
-                    {
-                        builder.Add(symbol.Name + "-" + arity.ToString(CultureInfo.InvariantCulture));
-                    }
-                    else
-                    {
-                        builder.Add(symbol.Name);
-                    }
-                }
-            }
-
-            INamedTypeSymbol containingType = symbol.ContainingType;
-
-            while (containingType != null)
-            {
-                int arity = containingType.Arity;
-
-                builder.Add((arity > 0) ? containingType.Name + "-" + arity.ToString(CultureInfo.InvariantCulture) : containingType.Name);
-
-                containingType = containingType.ContainingType;
-            }
-
-            INamespaceSymbol containingNamespace = symbol.ContainingNamespace;
-
-            if (containingNamespace != null)
-            {
-                if (containingNamespace.IsGlobalNamespace)
-                {
-                    if (symbol.Kind != SymbolKind.Namespace)
-                    {
-                        builder.Add(WellKnownNames.GlobalNamespaceName);
-                    }
-                }
-                else
-                {
-                    do
-                    {
-                        builder.Add(containingNamespace.Name);
-
-                        containingNamespace = containingNamespace.ContainingNamespace;
-                    }
-                    while (containingNamespace?.IsGlobalNamespace == false);
-                }
-            }
-
-            builder.Reverse();
-
-            return builder.ToImmutableArray();
-        }
-
         public override DocumentationUrlInfo CreateUrl(ISymbol symbol)
         {
-            ImmutableArray<string> parts = GetFolders(symbol);
-
-            return CreateUrl(parts);
-        }
-
-        public override DocumentationUrlInfo CreateUrl(ImmutableArray<string> folders)
-        {
-            if (!CanCreateUrl(folders))
+            if (!CanCreateUrl(symbol))
                 return default;
+
+            ImmutableArray<string> segments = DefaultUrlSegmentProvider.Hierarchical.GetSegments(symbol);
 
             const string baseUrl = "https://docs.microsoft.com/en-us/dotnet/api/";
 
             int capacity = baseUrl.Length;
 
-            foreach (string name in folders)
+            foreach (string name in segments)
                 capacity += name.Length;
 
-            capacity += folders.Length - 1;
+            capacity += segments.Length - 1;
 
             StringBuilder sb = StringBuilderCache.GetInstance(capacity);
 
             sb.Append(baseUrl);
 
-            sb.Append(folders[0].ToLowerInvariant());
+            sb.Append(segments[0].ToLowerInvariant());
 
-            for (int i = 1; i < folders.Length; i++)
+            for (int i = 1; i < segments.Length; i++)
             {
                 sb.Append(".");
-                sb.Append(folders[i].ToLowerInvariant());
+                sb.Append(segments[i].ToLowerInvariant());
             }
 
             return new DocumentationUrlInfo(StringBuilderCache.GetStringAndFree(sb), DocumentationUrlKind.External);
@@ -146,11 +51,6 @@ namespace Roslynator.Documentation
         public override bool CanCreateUrl(ISymbol symbol)
         {
             return IsWellKnownRootNamespace(symbol.GetRootNamespace()?.Name);
-        }
-
-        public override bool CanCreateUrl(ImmutableArray<string> folders)
-        {
-            return IsWellKnownRootNamespace(folders.FirstOrDefault());
         }
 
         private static bool IsWellKnownRootNamespace(string name)